Fixed compilation problem in Linux
Former-commit-id: 8aa926ed3434195cf94104b60b9c162ad110ec92 [formerly 828eb9d1e9bbf7f2907d2f58e7382371bf99a9a4]
Former-commit-id: a480686738517e3edb0e0ffeabd2d504db2d013e

+++ modules/ffmpeg/consumer/ffmpeg_consumer.cpp
@@ -791,7 +791,7 @@ public:
 
 	int64_t presentation_frame_age_millis() const override
 	{
-		return consumer_ ? consumer_->current_encoding_delay_ : 0;
+		return consumer_ ? static_cast<int64_t>(consumer_->current_encoding_delay_) : 0;
 	}
 
 	std::future<bool> send(core::const_frame frame) override
